requirements.txt
torch==1.7.1
python-telegram-bot==13.4.1
torchvision==0.8.2
numpy==1.20.1
opencv-python==4.5.1.48